// src/providers/Store/types/users.ts
export enum ActionTypeVerifyExistUser {
    VERIFY_EXIST_USER_START = 'VERIFY_EXIST_USER_START',
    VERIFY_EXIST_USER_SUCCESS = 'VERIFY_EXIST_USER_SUCCESS',
    VERIFY_EXIST_USER_ERROR = 'VERIFY_EXIST_USER_ERROR'
}

export enum ActionTypeSendCodeUser {
    SEND_CODE_USER_START = 'SEND_CODE_USER_START',
    SEND_CODE_USER_SUCCESS = 'SEND_CODE_USER_SUCCESS',
    SEND_CODE_USER_ERROR = 'SEND_CODE_USER_ERROR'
}

export enum ActionTypeVerifyCodeUser {
    VERIFY_CODE_USER_START = 'VERIFY_CODE_USER_START',
    VERIFY_CODE_USER_SUCCESS = 'VERIFY_CODE_USER_SUCCESS',
    VERIFY_CODE_USER_ERROR = 'VERIFY_CODE_USER_ERROR'
}

export enum ActionTypeSignUpTempUser {
    SIGNUP_TEMP_USER_START = 'SIGNUP_TEMP_USER_START',
    SIGNUP_TEMP_USER_SUCCESS = 'SIGNUP_TEMP_USER_SUCCESS',
    SIGNUP_TEMP_USER_ERROR = 'SIGNUP_TEMP_USER_ERROR'
}

export enum ActionTypeSignUpUser {
    SIGNUP_USER_START = 'SIGNUP_USER_START',
    SIGNUP_USER_SUCCESS = 'SIGNUP_USER_SUCCESS',
    SIGNUP_USER_ERROR = 'SIGNUP_USER_ERROR'
}

export enum ActionTypeVerifyEmailUser {
    VERIFY_EMAIL_USER_START = 'VERIFY_EMAIL_USER_START',
    VERIFY_EMAIL_USER_SUCCESS = 'VERIFY_EMAIL_USER_SUCCESS',
    VERIFY_EMAIL_USER_ERROR = 'VERIFY_EMAIL_USER_ERROR'
}

export enum ActionTypeProfileUser {
    PROFILE_USER_START = 'PROFILE_USER_START',
    PROFILE_USER_SUCCESS = 'PROFILE_USER_SUCCESS',
    PROFILE_USER_ERROR = 'PROFILE_USER_ERROR'
}

export enum ActionTypeSaveProfileUser {
    SAVE_PROFILE_USER_START = 'SAVE_PROFILE_USER_START',
    SAVE_PROFILE_USER_SUCCESS = 'SAVE_PROFILE_USER_SUCCESS',
    SAVE_PROFILE_USER_ERROR = 'SAVE_PROFILE_USER_ERROR'
}

export enum ActionTypeGenerateCodeProfileUser {
    GENERATE_CODE_PROFILE_USER_START = 'GENERATE_CODE_PROFILE_USER_START',
    GENERATE_CODE_PROFILE_USER_SUCCESS = 'GENERATE_CODE_PROFILE_USER_SUCCESS',
    GENERATE_CODE_PROFILE_USER_ERROR = 'GENERATE_CODE_PROFILE_USER_ERROR'
}

export enum ActionTypeVerifyCodeProfileUser {
    VERIFY_CODE_PROFILE_USER_START = 'VERIFY_CODE_PROFILE_USER_START',
    VERIFY_CODE_PROFILE_USER_SUCCESS = 'VERIFY_CODE_PROFILE_USER_SUCCESS',
    VERIFY_CODE_PROFILE_USER_ERROR = 'VERIFY_CODE_PROFILE_USER_ERROR'
}

export enum ActionTypeSaveContactProfileUser {
    SAVE_CONTACT_PROFILE_USER_START = 'SAVE_CONTACT_PROFILE_USER_START',
    SAVE_CONTACT_PROFILE_USER_SUCCESS = 'SAVE_CONTACT_PROFILE_USER_SUCCESS',
    SAVE_CONTACT_PROFILE_USER_ERROR = 'SAVE_CONTACT_PROFILE_USER_ERROR'
}

export enum ActionTypeSaveLocationProfileUser {
    SAVE_LOCATION_PROFILE_USER_START = 'SAVE_LOCATION_PROFILE_USER_START',
    SAVE_LOCATION_PROFILE_USER_SUCCESS = 'SAVE_LOCATION_PROFILE_USER_SUCCESS',
    SAVE_LOCATION_PROFILE_USER_ERROR = 'SAVE_LOCATION_PROFILE_USER_ERROR'
}

export enum ActionTypeDeleteAccountUser {
    DELETE_ACCOUNT_USER_START = 'DELETE_ACCOUNT_USER_START',
    DELETE_ACCOUNT_USER_SUCCESS = 'DELETE_ACCOUNT_USER_SUCCESS',
    DELETE_ACCOUNT_USER_ERROR = 'DELETE_ACCOUNT_USER_ERROR'
}